continue to fix middleware token.js
